Issue #12: Move routeStore into own folder
We need to respect the need to have the staeMachineDefinition in its own file
At this point there are more non-store files in the store/ folder than actual store files
I doubt we'll need another store but for the reason each store should get its own folder
Use an index.js file in each folder to respect existing store-related imports
