+++ src/stores/router/index.js
@@ -1,0 +1,2 @@
+import { getStore, transition } from './router';
+export { getStore, transition };
